+++ app/contracts/teams/players/update_contract.rb
@@ -9,7 +9,6 @@ module Teams
         optional(:active)
         optional(:price_cents).filled(:integer)
         optional(:shirt_number_string).filled(:string)
-        optional(:form).filled(:float)
       end
     end
   end