+++ app/controllers/seasons/players_controller.rb
@@ -17,7 +17,7 @@ module Seasons
 
     def find_season_players
       @season_players =
-        Season.active.find_by!(uuid: params[:season_id]).active_teams_players
+        Season.active.find_by!(uuid: params[:season_id]).players_seasons
     end
 
     def find_season_player
@@ -26,21 +26,23 @@ module Seasons
 
     def season_players_json_response
       Rails.cache.fetch(
-        ['seasons_players_index_v1', @season_players.maximum(:updated_at)],
+        ['seasons_players_index_v2', @season_players.maximum(:updated_at)],
         expires_in: 6.hours,
         race_condition_ttl: 10.seconds
       ) do
-        Teams::PlayerSerializer.new(@season_players.includes(:player, seasons_team: :team)).serializable_hash
+        ::Players::SeasonSerializer.new(
+          @season_players.includes(:player, active_teams_player: [seasons_team: :team])
+        ).serializable_hash
       end
     end
 
     def season_player_json_response
       Rails.cache.fetch(
-        ['seasons_players_show_v1', @season_player.id, @season_player.updated_at],
+        ['seasons_players_show_v2', @season_player.id, @season_player.updated_at],
         expires_in: 12.hours,
         race_condition_ttl: 10.seconds
       ) do
-        Controllers::Seasons::Players::ShowSerializer.new(@season_player).serializable_hash
+        ::Controllers::Seasons::Players::ShowSerializer.new(@season_player).serializable_hash
       end
     end
   end