+++ app/controllers/weeks/transfers_controller.rb
@@ -12,7 +12,7 @@ module Weeks
 
     def json_response
       Rails.cache.fetch(
-        ['weeks_transfers_index_v1', @week.transfers.maximum(:updated_at)],
+        ['weeks_transfers_index_v2', @week.transfers.maximum(:updated_at)],
         expires_in: 24.hours,
         race_condition_ttl: 10.seconds
       ) do
@@ -41,7 +41,7 @@ module Weeks
 
     def find_transfers(grouped_teams_players_ids)
       transfers = grouped_teams_players_ids.sort_by { |_k, v| -v }.first(10).to_h
-      teams_players = Teams::Player.where(id: transfers.keys).includes(:player, seasons_team: :team)
+      teams_players = Teams::Player.where(id: transfers.keys).includes(:player, :players_season, seasons_team: :team)
       transfers.map do |teams_player_id, amount|
         [
           Controllers::Weeks::Transfers::IndexSerializer