+++ app/javascript/components/TransfersStatus/TransfersStatus.jsx
@@ -55,7 +55,7 @@ export const TransfersStatus = ({ weekUuid, seasonUuid, sportKind }) => {
             <tr key={index}>
               <td
                 className="cursor-pointer"
-                onClick={() => setPlayerUuid(item[0].data.attributes.uuid)}
+                onClick={() => setPlayerUuid(item[0].data.attributes.players_season.uuid)}
               >i</td>
               <td>{localizeValue(item[0].data.attributes.player.name).split(' ')[0]}</td>
               <td>{localizeValue(sportPositions[item[0].data.attributes.player.position_kind].name)}</td>