+++ app/jobs/players/seasons/mass_update_job.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
 module Players
-  module Statistic
-    class UpdateJob < ApplicationJob
+  module Seasons
+    class MassUpdateJob < ApplicationJob
       queue_as :default
 
       def perform(season_id:, player_ids:)
-        Players::Statistic::UpdateService.call(
+        Players::Seasons::MassUpdateService.call(
           season_id: season_id,
           player_ids: player_ids
         )