+++ app/models/teams/player.rb
@@ -10,6 +10,7 @@ module Teams
 
     belongs_to :seasons_team, class_name: '::Seasons::Team'
     belongs_to :player, class_name: '::Player'
+    belongs_to :players_season, class_name: '::Players::Season'
 
     has_many :games_players, class_name: '::Games::Player', foreign_key: :teams_player_id, dependent: :destroy
     has_many :games, through: :games_players