+++ app/services/teams/players/price/change_service.rb
@@ -19,13 +19,14 @@ module Teams
           teams_players =
             Teams::Player
               .where(id: transfers_data.keys)
-              .hashable_pluck(:id, :price_cents, :seasons_team_id, :player_id)
+              .hashable_pluck(:id, :price_cents, :seasons_team_id, :player_id, :players_season_id)
               .map do |teams_player|
                 {
                   id: teams_player[:id],
                   price_cents: teams_player[:price_cents] + price_modified(teams_player[:id]),
                   seasons_team_id: teams_player[:seasons_team_id],
-                  player_id: teams_player[:player_id]
+                  player_id: teams_player[:player_id],
+                  players_season_id: teams_player[:players_season_id]
                 }
               end
           # commento: teams_players.price_cents