+++ spec/services/weeks/start_service_spec.rb
@@ -5,7 +5,6 @@ describe Weeks::StartService, type: :service do
     described_class
       .new(
         price_change_service: price_change_service,
-        form_change_service: form_change_service,
         cup_create_service: cup_create_service,
         cups_pairs_generate_service: cups_pairs_generate_service,
         generate_week_position: generate_week_position
@@ -14,14 +13,12 @@ describe Weeks::StartService, type: :service do
   }
 
   let(:price_change_service) { double }
-  let(:form_change_service) { double }
   let(:cup_create_service) { double }
   let(:cups_pairs_generate_service) { double }
   let(:generate_week_position) { 2 }
 
   before do
     allow(price_change_service).to receive(:call)
-    allow(form_change_service).to receive(:call)
     allow(cup_create_service).to receive(:call)
     allow(cups_pairs_generate_service).to receive(:call)
   end
@@ -38,7 +35,6 @@ describe Weeks::StartService, type: :service do
 
       expect(week).not_to have_received(:update)
       expect(price_change_service).not_to have_received(:call)
-      expect(form_change_service).not_to have_received(:call)
       expect(cup_create_service).not_to have_received(:call)
       expect(cups_pairs_generate_service).not_to have_received(:call)
       expect(service.success?).to be_truthy
@@ -54,7 +50,6 @@ describe Weeks::StartService, type: :service do
 
         expect(week.reload.status).not_to eq Week::ACTIVE
         expect(price_change_service).not_to have_received(:call)
-        expect(form_change_service).not_to have_received(:call)
         expect(cup_create_service).not_to have_received(:call)
         expect(cups_pairs_generate_service).not_to have_received(:call)
         expect(service.success?).to be_truthy
@@ -64,17 +59,18 @@ describe Weeks::StartService, type: :service do
     context 'for coming week' do
       let!(:previous_week) { create :week, status: Week::FINISHED, position: 1 }
       let!(:week) { create :week, status: Week::COMING, position: 2, season: previous_week.season }
-      let!(:game1) { create :game, week: previous_week }
-      let!(:game2) { create :game, week: week }
       let!(:fantasy_league) { create :fantasy_league, leagueable: previous_week.season, season: previous_week.season }
 
-      before { create :lineup, week: week }
+      before do
+        create :game, week: previous_week
+        create :game, week: week
+        create :lineup, week: week
+      end
 
       it 'updates week, calls change_service and succeed', :aggregate_failures do
         expect { service_call }.to change(week.fantasy_leagues, :count).by(1)
         expect(week.reload.status).to eq Week::ACTIVE
         expect(price_change_service).to have_received(:call).with(week: week)
-        expect(form_change_service).to have_received(:call).with(games_ids: [game1.id, game2.id])
         expect(cup_create_service).to have_received(:call).with(fantasy_league: fantasy_league)
         expect(cups_pairs_generate_service).not_to have_received(:call)
         expect(service_call.success?).to be_truthy
@@ -87,7 +83,6 @@ describe Weeks::StartService, type: :service do
           service = service_call
 
           expect(price_change_service).to have_received(:call).with(week: week)
-          expect(form_change_service).to have_received(:call).with(games_ids: [game1.id, game2.id])
           expect(cup_create_service).not_to have_received(:call)
           expect(cups_pairs_generate_service).not_to have_received(:call)
           expect(service.success?).to be_truthy
@@ -101,7 +96,6 @@ describe Weeks::StartService, type: :service do
           service = service_call
 
           expect(price_change_service).to have_received(:call).with(week: week)
-          expect(form_change_service).to have_received(:call).with(games_ids: [game1.id, game2.id])
           expect(cup_create_service).not_to have_received(:call)
           expect(cups_pairs_generate_service).to have_received(:call).with(week: week)
           expect(service.success?).to be_truthy